examples/databricks_DBRX_website_bot/requirements.txt
llama-index
llama-index-llms-databricks
llama-index-embeddings-huggingface
llama-index-readers-web
llama-index-vector-stores-lancedb
diffusers
mlx>=0.11
huggingface-hub
regex
numpy
tqdm
Pillow
streamlit